fix enumeration with multiple devs are possible
cv2.VideoCapture.read() returns a tuple, so checking None here would always return True - a bad camera would return (False, None). Sometimes this works, but it depends on the order of enumeration which is not guaranteed. So we first sort the IDs to get the lowest index (usually the right one), then we properly check that the capture succeeded or not.