Added: get_serial_device_list() which finds and lists all com ports and associated serial number with Boson attached Added find_video_devices_win32() which lists all the FLIR Video devices and their ID.
The code is changed to use these functions in stead of the find_serial_device() and find_video_device() uses the find_video_devices_win32() for win32 platform.
Only the first found device is actually used in the Boson object.

The test_boson.py is also updated to test the new functions - it is however probably not the right way to do this. So update here is necessary.

To fully exploit the benefit of these new functions in Boson class there need to be a bit of a rewrite in the rest of the class. As this will touch on the interfaces so I will not do this.

The code is seriously lacking a feature to correlate which COM port is attached to which device ID.