Update frame offset calculation
In the previous code, the chunksize variable always come from the current position minus the previous position. In each tuple, the first element is the current position, whereas the second element is the previous chunk size. This means that there will always be a mismatch in each tuple.
Only fixing up the first chunk size will NOT work correctly, unless all chunks have the SAME size. In which case it is simply updating the first 0 to be the actual chunk size.
The updated code should solve this issue.